Allow iteration over parent HeaderSchemas
In the original implementation, the `HeaderSchema` parent classes were not
included in `__bases__` of a `BinaryTable`.  This made it impossible to
iterate over groups (classes) of headers, for documentation purposes,
since all you could access was the merged list of headers.

This makes some breaking changes:

- removes complex mechanism to "create" a new `HeaderSchema` class for
`BinaryTables`, whose purpose was to ensure that the `HeaderSchemas`
automatically inherited from `BinaryTableHeader`. Now, we just require (and
check) that `BinaryTableHeader` is a parent class. Less automatic, but more
explicit.
- adds a helper class method `_header_schemas()` that returns a filtered
list of bases that inherit from `HeaderSchema`