Support string columns as Char columns
Changes dtype check to look only for the base dtype in the case of
columns with numpy dtypes 'S' or 'U'. Oddly, astropy.io.fits seems to do
some type conversion when accessing a column of a FITS_rec, so this
should work with HDUs, but perhaps not with bare numpy recarrays or
astropy tables, which do not convert 'S' to 'U' on column access.

Also changes slightly the logic of the unit check, to only check if a
unit was specified, since string columns are not convertable to unit
quantities.